+++ resources/config.py
@@ -23,6 +23,7 @@ from schema.config import (
 from . import logger, requires_authorization, deprecated
 from .object import (
     list_objects, get_object_creation_params, get_object,
+    get_legacy_form_options,
     ObjectResource, ObjectsResource
 )
 
@@ -243,7 +244,7 @@ class ConfigResource(ObjectResource):
                 description: Object exists yet but has different type
         """
         schema = ConfigLegacyCreateRequestSchema()
-        params = schema.loads(request.get_data(as_text=True))
+        params = schema.load(get_legacy_form_options())
 
         if params and params.errors:
             return {"errors": params.errors}, 400
